update APIv2
Took 1 hour 14 minutes

+++ APIv2/database/ban_list.py
@@ -7,7 +7,7 @@ from .db_session import SqlAlchemyBase
 class Ban_list(SqlAlchemyBase):
     __tablename__ = "Ban_list"
     id = sqlalchemy.Column(sqlalchemy.Integer, primary_key=True, autoincrement=True)
-    tg_id = sqlalchemy.Column(sqlalchemy.Integer, nullable=False)
+    tg_id = sqlalchemy.Column(sqlalchemy.BigInteger, nullable=False)
     class_id = sqlalchemy.Column(
         sqlalchemy.Integer, sqlalchemy.ForeignKey("classes.id"), nullable=False
     )
